Revert 'fix' for template order. Update test cases to reflect expected inference
